backend/requirements.txt
fastapi
uvicorn
openai==0.28
chromadb
python-dotenv
sentence-transformers
python-multipart
